// client/app/models/Weather.ts
/**
 * Created by Josh Chan on 02-12-2016.
 */

export class Weather {
    _id: string;
    year: Number;
    month: Number;
    day: Number;
    temp: Number;
    humidity: Number;        
    cloud: Number;              
    rainfall: Number;              
    rainfallDescEN: string;
    rainfallDescZH: string;
    sunshine: Number;
}