// client/app/models/ResultWeather.ts
/**
 * Created by Josh Chan on 02-12-2016.
 */

export class ResultWeather {
    year: Number;
    month: Number;
    day: Number;
    weekDay: Number;
    rainYear: Number;
    drizzle: Number;
    light: Number;
    moderate: Number;
    heavy: Number;
    violent: Number;
    torrential: Number;
    temp: Number;
    humidity: Number;
    cloud: Number;
    sunshine: Number;
}